Breaking down lib.rs into separate entities as proposed cleanup
Match the readme.md example


Preserving original API


Few missed bits


Make cargo fmt happy
